:pencil: Updated the readme with more information.
