+++ Assets/notes.md
@@ -1,0 +1,1 @@
+The assets folder is used, so that load the images in the wiki.
