--- Static/index.html
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="style.css">
-    <title>Customize profile</title>
-</head>
-
-<body>
-    <header>
-        <nav>
-            <ul>
-                <li><a href="#"><img src="img/Arrow_Back.png" alt="Go back to the overview page"></a></li>
-                <li>
-                    <a href="#"><img src="img/chat.png" alt="Go to your private messages"></a>
-                </li>
-            </ul>
-        </nav>
-    </header>
-    <main>
-        <h1>John</h1>
-        <img id="ProfilePlaceholder" src="img/placeholder.png" alt="Go back to the overview page">
-        <a href="#">Edit photo</a>
-        <form action="#">
-            <fieldset>
-                <legend>My details</legend>
-                <input placeholder="About me" type="text" name="Bio" id="bio">
-                <input placeholder="Education/Proffesion" type="text" name="Education" id="education">
-                <input placeholder="Email" type="email" name="Email" id="email">
-                <input placeholder="Date of birth" type="date" name="DoB" id="dob">
-
-                <input type="submit" value="Save changes">
-            </fieldset>
-        </form>
-        <a id="premium" href="#">Go premium for €5,99 a month.</a><br>
-        <a id="SettingsLink" href="moresettings.html">More settings</a>
-    </main>
-</body>
-
-</html>
