--- Static/moresettings.html
@@ -1,63 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="style.css">
-    <title>More settings</title>
-</head>
-
-<body>
-    <header>
-        <nav>
-            <ul>
-                <li><a href="index.html"><img src="img/Arrow_Back.png" alt="Go back to the overview page"></a></li>
-                <li>
-                    <a href="#"><img src="img/chat.png" alt="Go to your private messages"></a>
-                </li>
-            </ul>
-        </nav>
-    </header>
-    <main>
-        <a href="index.html">Profile settings</a>
-        <form class="moreSettings" action="#">
-            <fieldset>
-                <legend>More settings</legend>
-                <h2>Push notifications:</h2>
-                Off <input type="checkbox" name="notifications" id="notifications"> On<br>
-                <h2>My gender:</h2> Male <input type="radio" name="gender" id="genderMale"><br>
-                Female <input type="radio" name="gender" id="genderFemale"><br>
-                Other <input type="radio" name="gender" id="genderOther"><br>
-                <h2>Interested in:</h2>
-                Male <input type="checkbox" name="interest" id="interestMale"><br>
-                Female <input type="checkbox" name="interest" id="interestFemale"><br>
-                Other <input type="checkbox" name="interest" id="interestFemale"><br>
-                <h2>Prefered zodiac's:</h2>
-                <div id="zodiacGrid">
-                    <ul>
-                        <li><input type="checkbox" name="interest" id="interestMale"> Aquarius</li>
-                        <li><input type="checkbox" name="interest" id="interestMale"> Pisces</li>
-                        <li><input type="checkbox" name="interest" id="interestMale"> Aries</li>
-                        <li><input type="checkbox" name="interest" id="interestMale"> Taurus</li>
-                        <li><input type="checkbox" name="interest" id="interestMale"> Gemini</li>
-                        <li><input type="checkbox" name="interest" id="interestMale"> Cancer</li>
-                        <li><input type="checkbox" name="interest" id="interestMale"> Leo</li>
-                        <li><input type="checkbox" name="interest" id="interestMale"> Virgo</li>
-                        <li><input type="checkbox" name="interest" id="interestMale"> Libra</li>
-                        <li><input type="checkbox" name="interest" id="interestMale"> Scorpio</li>
-                        <li><input type="checkbox" name="interest" id="interestMale"> Sagittarius</li>
-                        <li><input type="checkbox" name="interest" id="interestMale"> Capricorn</li>
-                    </ul>
-                </div>
-                <h2>Min age:</h2> <input type="range" list="tickmarks" min="0" max="100" step="5"><br>
-                <h2>Max age:</h2> <input type="range" list="tickmarks" min="0" max="100" step="5"><br>
-                <h2>MaximumDistance:</h2> <input type="range" list="tickmarks" min="0" max="100" step="5"><br>
-                <input type="submit" value="Save changes">
-            </fieldset>
-        </form>
-    </main>
-</body>
-
-</html>
