--- Static/style.css
@@ -1,132 +1,0 @@
-html,
-body {
-    font-family: 'Roboto', sans-serif;
-    margin: 0;
-    font-size: 1em;
-
-}
-
-header {
-    height: 3.125em;
-    background-color: #343557;
-    width: 100%;
-    margin-bottom: 3%;
-    position: relative;
-}
-
-#premium {
-    background-color: #448ad3;
-    border: 0;
-    font-weight: bold;
-    font-size: 1rem;
-    color: #fff;
-    padding: 0.5625em 0.9375em;
-    border-radius: 10px;
-    font-style: normal;
-    text-decoration: none;
-    margin-top: 2%;
-}
-
-#SettingsLink {
-    display: block;
-    margin-top: 4%;
-}
-
-nav img {
-    width: 1.3em;
-    display: block;
-    margin: 0 auto;
-}
-
-nav ul {
-    display: flex;
-    height: 3.125em;
-    list-style: none;
-    padding: 0;
-    margin: 0;
-    position: relative;
-}
-
-nav li {
-    flex: 1;
-    align-items: center;
-
-    display: flex;
-}
-
-nav ul :nth-child(1) {
-    justify-content: flex-start;
-    margin-left: 4%;
-}
-
-nav ul :nth-child(2) {
-    justify-content: flex-end;
-    margin-right: 4%;
-}
-
-h1 {
-    text-align: center;
-    color: #343557;
-}
-
-main {
-    width: 100%;
-    text-align: center;
-}
-
-#ProfilePlaceholder {
-    width: 40%;
-    margin: 0 auto;
-    display: block;
-}
-
-main form fieldset {
-    border: none;
-    margin-top: 5%;
-
-}
-
-main form legend {
-    font-size: 1.5em;
-    color: #343557;
-}
-
-main form input {
-    display: block;
-    margin: auto;
-    border: none;
-    font-size: 1rem;
-    text-align: center;
-    margin-bottom: 2%;
-    border-bottom: 1px solid #343557;
-    color: #989899;
-    text-align: center;
-    font-style: italic;
-}
-
-main form input[type="submit"] {
-    background-color: #aa3b6a;
-    border: 0;
-    font-weight: bold;
-    font-size: 1rem;
-    color: #fff;
-    padding: 0.5625em 0.9375em;
-    border-radius: 10px;
-    margin-top: 4%;
-    font-style: normal;
-}
-
-.moreSettings {
-    margin-top: 5%;
-}
-
-.moreSettings input {
-    display: inline;
-}
-
-#zodiacGrid ul {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    text-align: left;
-    list-style: none;
-}
