+++ app.json
@@ -1,0 +1,6 @@
+{
+    "name": "Zodimaniac",
+    "description": "Dating platform for people intrested in astrology",
+    "repository": "https://github.com/Ramon96/Project-tech-team8",
+    "keywords": ["Node, Express, mongodb"]
+  }
