--- routes/app.json
@@ -1,6 +1,0 @@
-{
-    "name": "Zodimaniac",
-    "description": "Edit profile feature",
-    "repository": "https://github.com/Ramon96/project-tech",
-    "keywords": ["Node, Express, mongodb"]
-  }
