+++ routes/signup.js
@@ -1,0 +1,8 @@
+var express = require('express');
+var router = express.Router();
+
+router.get('/', function(req, res){
+    res.render('signup', {message: "hello"});
+});
+
+module.exports = router;
