+++ views/editprofile.hbs
@@ -1,5 +1,5 @@
 <h1>Edit profile</h1>
-<form action="/editprofile/update" method="POST">
+<form action="editprofile/update" method="POST">
     <label for="aboutme">About me</label>
     <textarea name="aboutme" id="aboutme">{{aboutme}}</textarea>
