+++ views/partials/header.hbs
@@ -1,7 +1,7 @@
     <header>
         <nav>
             <ul>
-                <li><a href="/"><img src="/images/Arrow Back.png" alt="Go back to the overview page"></a></li>
+                <li><a href="/edit"><img src="/images/Arrow Back.png" alt="Go back to the overview page"></a></li>
             </ul>
         </nav>
     </header>
