+++ views/settings.hbs
@@ -1,5 +1,5 @@
 <h1>Settings</h1>
-<form action="/settings/update" method="POST">
+<form action="/edit/settings/update" method="POST">
     <label for="username">Username</label>
     {{#if usernameIsTaken}}
     <p class="warning">{{usernameTaken}}</p>
